Event handler life cycle management
* Support new code deploys
* Subscription to on eventing role nodes to get notified
  on app code deploy or config changes
* Retiring couple of settings from deployment config, it wasn't
  right place for them
* Expose couple of endpoints to verify/configure app specific
  settings(workerCount, statsTicker)
* Address goroutine leaks on app redeploys
* First cut of topology aware eventing mds role

Reviewed-on: http://review.couchbase.org/74075
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ consumer/v8_consumer.go
@@ -30,6 +30,7 @@ func New(p common.EventingProducer, app *common.AppConfig, vbnos []uint16, bucke
 		producer:                  p,
 		signalConnectedCh:         make(chan bool),
 		statsTicker:               time.NewTicker(StatsTickInterval),
+		stopControlRoutineCh:      make(chan bool),
 		tcpPort:                   tcpPort,
 		vbFlogChan:                make(chan *vbFlogEntry),
 		vbnos:                     vbnos,
@@ -115,9 +116,12 @@ func (c *Consumer) Stop() {
 
 	c.producer.CleanupDeadConsumer(c)
 
-	c.cmd.Process.Kill()
+	if c.osPid != 0 {
+		c.cmd.Process.Kill()
+	}
 
 	c.statsTicker.Stop()
+	c.stopControlRoutineCh <- true
 	c.stopCheckpointingCh <- true
 	c.gracefulShutdownChan <- true
 	c.dcpFeed.Close()
